// packages/trading-core/src/solvers/types/generated/rasa-solver.ts
export interface SymbolsContract {
  price_precision: number;
  quantity_precision: number;
  name: string;
  symbol: string;
  asset: string;
  symbol_id: number;
  is_valid: boolean;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  min_acceptable_quote_value: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  min_acceptable_portion_lf: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  trading_fee: string;
  max_leverage: number;
  max_notional_value: number;
  rfq_allowed: boolean;
  hedger_fee_open: string;
  hedger_fee_close: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  max_funding_rate: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  min_notional_value: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  max_quantity: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  lot_size: string;
}